requirements.txt
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
click==8.1.7
Flask==2.2.5
greenlet==3.0.3
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.21.6
pandas==1.3.5
psycopg2-binary==2.9.9
pymssql==2.2.11
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
